Move node-ctrl crate into node crate and rename node-ctrl-proto into node-services
This commit moves the node-ctrl crate into the node crate under the module
server. Moreover, this commit renames the node-ctrl-proto crate into
node-services.
