+++ Cargo.toml
@@ -46,8 +46,7 @@ restate-meta = { path = "crates/meta" }
 restate-meta-rest-model = { path = "crates/meta-rest-model" }
 restate-network = { path = "crates/network" }
 restate-node = { path = "crates/node" }
-restate-node-ctrl = { path = "crates/node-ctrl" }
-restate-node-ctrl-proto = { path = "crates/node-ctrl-proto" }
+restate-node-services = { path = "crates/node-services" }
 restate-pb = { path = "crates/pb" }
 restate-queue = { path = "crates/queue" }
 restate-schema-api = { path = "crates/schema-api" }
