--- crates/node-ctrl/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "restate-node-ctrl"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-publish = false
-
-[features]
-default = []
-options_schema = ["dep:schemars"]
-
-[dependencies]
-restate-bifrost = { workspace = true }
-restate-cluster-controller = { workspace = true }
-restate-errors = { workspace = true }
-restate-node-ctrl-proto = { workspace = true }
-restate-storage-rocksdb = { workspace = true }
-restate-types = { workspace = true }
-
-async-trait = { workspace = true }
-axum = { workspace = true }
-codederror = { workspace = true }
-derive_builder = { workspace = true }
-drain = { workspace = true }
-futures = { workspace = true }
-http = { workspace = true }
-humantime = { workspace = true }
-hyper = { workspace = true, features = ["full"] }
-metrics = { workspace = true }
-metrics-exporter-prometheus = { version = "0.13", default-features = false, features = ["async-runtime"] }
-metrics-tracing-context = { version = "0.15.0" }
-metrics-util = { version = "0.16.0" }
-rocksdb = { workspace = true }
-schemars = { workspace = true, optional = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-serde_with = { workspace = true }
-thiserror = { workspace = true }
-tonic = { workspace = true, features = ["codegen", "prost", "transport"] }
-tonic-reflection = { workspace = true }
-tower = { workspace = true, features = ["load-shed", "limit"] }
-tower-http = { workspace = true, features = ["trace"] }
-tracing = { workspace = true }
