+++ crates/node-services/proto/cluster_controller.proto
