+++ crates/node/src/server/handler/cluster_controller.rs
@@ -8,8 +8,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use restate_node_ctrl_proto::cluster_controller::cluster_controller_server::ClusterController;
-use restate_node_ctrl_proto::cluster_controller::{AttachmentRequest, AttachmentResponse};
+use restate_node_services::cluster_controller::cluster_controller_server::ClusterController;
+use restate_node_services::cluster_controller::{AttachmentRequest, AttachmentResponse};
 use restate_types::NodeId;
 use tonic::{async_trait, Request, Response, Status};
 use tracing::debug;
